+++ NEWS.md
@@ -1,4 +1,7 @@
 # dimensio 0.8.1.9000
+## New classes and methods
+* Add `pcoa()` to compute principal coordinates analysis.
+
 ## Enhancements
 * Allow to display extra qualitative and quantitative variables in `viz_rows()`/`viz_individuals()` and `viz_columns()`/`viz_variables()`.